+++ packages/verified-fetch/test/ipns-record.spec.ts
@@ -1,25 +1,26 @@
 import { dagCbor } from '@helia/dag-cbor'
-import { ipns } from '@helia/ipns'
 import { generateKeyPair } from '@libp2p/crypto/keys'
 import { stop } from '@libp2p/interface'
 import { peerIdFromPrivateKey } from '@libp2p/peer-id'
 import { expect } from 'aegir/chai'
-import { marshalIPNSRecord, unmarshalIPNSRecord } from 'ipns'
+import { createIPNSRecord, marshalIPNSRecord, unmarshalIPNSRecord } from 'ipns'
+import { stubInterface } from 'sinon-ts'
 import { VerifiedFetch } from '../src/verified-fetch.js'
 import { createHelia } from './fixtures/create-offline-helia.js'
 import type { Helia } from '@helia/interface'
-import type { IPNS } from '@helia/ipns'
+import type { IPNSResolver } from '@helia/ipns'
+import type { StubbedInstance } from 'sinon-ts'
 
 describe('ipns records', () => {
   let helia: Helia
-  let name: IPNS
+  let ipnsResolver: StubbedInstance<IPNSResolver>
   let verifiedFetch: VerifiedFetch
 
   beforeEach(async () => {
     helia = await createHelia()
-    name = ipns(helia)
-    verifiedFetch = new VerifiedFetch({
-      helia
+    ipnsResolver = stubInterface()
+    verifiedFetch = new VerifiedFetch(helia, {
+      ipnsResolver
     })
   })
 
@@ -34,9 +35,16 @@ describe('ipns records', () => {
     const c = dagCbor(helia)
     const cid = await c.add(obj)
 
-    const key = await generateKeyPair('Ed25519')
-    const peerId = peerIdFromPrivateKey(key)
-    const record = await name.publish(key, cid)
+    const privateKey = await generateKeyPair('Ed25519')
+    const record = await createIPNSRecord(privateKey, cid, 1, 10_000)
+    const peerId = peerIdFromPrivateKey(privateKey)
+
+    ipnsResolver.resolve.withArgs(peerId).resolves({
+      cid,
+      record
+    })
+
+    const marshaledRecord = marshalIPNSRecord(record)
 
     const resp = await verifiedFetch.fetch(`ipns://${peerId}`, {
       headers: {
@@ -45,12 +53,42 @@ describe('ipns records', () => {
     })
     expect(resp.status).to.equal(200)
     expect(resp.headers.get('content-type')).to.equal('application/vnd.ipfs.ipns-record')
+    expect(resp.headers.get('content-length')).to.equal(marshaledRecord.byteLength.toString())
+    expect(resp.headers.get('x-ipfs-roots')).to.equal(cid.toV1().toString())
+    expect(resp.headers.get('content-disposition')).to.equal(`attachment; filename="${peerId}.bin"`)
 
     const buf = new Uint8Array(await resp.arrayBuffer())
     expect(marshalIPNSRecord(record)).to.equalBytes(buf)
 
     const output = unmarshalIPNSRecord(buf)
     expect(output.value).to.deep.equal(`/ipfs/${cid}`)
+  })
+
+  it('should override filename when fetching a raw IPNS record', async () => {
+    const obj = {
+      hello: 'world'
+    }
+    const c = dagCbor(helia)
+    const cid = await c.add(obj)
+
+    const privateKey = await generateKeyPair('Ed25519')
+    const record = await createIPNSRecord(privateKey, cid, 1, 10_000)
+    const peerId = peerIdFromPrivateKey(privateKey)
+
+    ipnsResolver.resolve.withArgs(peerId).resolves({
+      cid,
+      record
+    })
+
+    const filename = 'foo.bin'
+
+    const resp = await verifiedFetch.fetch(`http://${peerId}.ipns.local?filename=${filename}`, {
+      headers: {
+        accept: 'application/vnd.ipfs.ipns-record'
+      }
+    })
+    expect(resp.status).to.equal(200)
+    expect(resp.headers.get('content-disposition')).to.equal(`attachment; filename="${filename}"`)
   })
 
   it('should reject a request for non-IPNS url', async () => {
@@ -71,18 +109,8 @@ describe('ipns records', () => {
     expect(resp.status).to.equal(400)
   })
 
-  it('should reject a request for a url with a path component', async () => {
-    const obj = {
-      hello: 'world'
-    }
-    const c = dagCbor(helia)
-    const cid = await c.add(obj)
-
-    const key = await generateKeyPair('Ed25519')
-    const peerId = peerIdFromPrivateKey(key)
-    await name.publish(key, cid)
-
-    const resp = await verifiedFetch.fetch(`ipns://${peerId}/hello`, {
+  it('should reject a request for an IPNS url with a path component', async () => {
+    const resp = await verifiedFetch.fetch('ipns://QmbxpRxwKXxnJQjnPqm1kzDJSJ8YgkLxH23mcZURwPHjGv/hello', {
       headers: {
         accept: 'application/vnd.ipfs.ipns-record'
       }